requirements.txt
numpy==1.26.4
pandas==2.2.2
xarray==2024.9.0
rioxarray==0.19.0
scipy==1.13.1
matplotlib==3.8.4
cartopy==0.24.0
cmcrameri==0.0.0
geopandas==1.1.1
PyYAML==6.0.1
